backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
PyMySQL==1.1.0